+++ apps/web/app/lib/components/Onboarding/PersonalInfoStep.tsx
@@ -4,7 +4,7 @@ import { Button } from "@repo/ui/components/ui/button";
 import { Input } from "@repo/ui/components/ui/input";
 import { RadioGroup, RadioGroupItem } from "@repo/ui/components/ui/radio-group";
 import { User, ArrowLeft, ArrowRight } from "lucide-react";
-import type { OnboardingData } from "../../signup/onboarding";
+import type { OnboardingData } from "@/app/signup/onboarding";
 
 interface PersonalInfoStepProps {
   onNext: (data: Partial<OnboardingData>) => void;
@@ -94,7 +94,7 @@ export function PersonalInfoStep({
         </Button>
         <Button type="submit" className="bg-black text-white hover:bg-gray-800">
           Next
-          <ArrowRight/>
+          <ArrowRight />
         </Button>
       </div>
     </form>
