+++ apps/web/app/mentorship/find-mentorship/mentee-list.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from "react";
 import { DataTable } from "./data-table"
-import { SortOption } from "@/app/lib/components/sort-mentors";
+import { SortOption } from "@/app/lib/components/MentorshipTable/sort-mentors";
 import { User } from "@/app/lib/types";
 import { Mentee, menteeColumns } from "./mentee-columns";
