+++ apps/web/app/mentorship/find-mentorship/mentor-list.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from "react";
 import { DataTable } from "./data-table"
-import { SortOption } from "@/app/lib/components/sort-mentors";
+import { SortOption } from "@/app/lib/components/MentorshipTable/sort-mentors";
 import { Mentor, mentorColumns } from "./mentor-columns";
 import { User } from "@/app/lib/types";
