+++ apps/web/app/mentorship/logs/page.tsx
@@ -1,7 +1,7 @@
-import LogsSidebar from "../../lib/components/logs-sidebar";
-import NewInteractionForm from "../../lib/components/new-interaction-form";
-import InteractionDetails from "../../lib/components/interaction-details";
-import LogsContent from "../../lib/components/logs-content";
+import LogsSidebar from "../../lib/components/InteractionLogs/logs-sidebar";
+import NewInteractionForm from "../../lib/components/InteractionLogs/new-interaction-form";
+import InteractionDetails from "../../lib/components/InteractionLogs/interaction-details";
+import LogsContent from "../../lib/components/InteractionLogs/logs-content";
 
 interface Interaction {
     _id: string;
@@ -14,19 +14,19 @@ interface Interaction {
 
 export default async function Logs({ searchParams }: { searchParams: { [key: string]: string | string[] | undefined } }) {
 
-  if (!searchParams) return null;
-  const showForm = searchParams['new'] === 'true';
-  const selectedId = searchParams['id'];
+    if (!searchParams) return null;
+    const showForm = searchParams['new'] === 'true';
+    const selectedId = searchParams['id'];
 
-  const allLogsResponse = await fetch("http://localhost:2999/mentorship/logs", {
-      cache: "no-store",
-  });
+    const allLogsResponse = await fetch("http://localhost:2999/mentorship/logs", {
+        cache: "no-store",
+    });
 
-  const allLogs: { data: Interaction[] } = await allLogsResponse.json();
+    const allLogs: { data: Interaction[] } = await allLogsResponse.json();
 
-  const selectedInteraction = allLogs.data.find(
-    (interaction: Interaction) => interaction._id === selectedId
-  );
+    const selectedInteraction = allLogs.data.find(
+        (interaction: Interaction) => interaction._id === selectedId
+    );
 
     return (
         <div className="flex h-[calc(100vh-68px)] overflow-hidden">
